quelques corrections

+++ Services/IPaiementService.cs
@@ -8,8 +8,8 @@ namespace GestionCoutureApp.Services
         List<Paiement> ObtenirParCommande(int idCommande);
         void Ajouter(Paiement paiement, int idOperateur, string nomOperateur);
         void Annuler(int idPaiement, string motif, string nomAnnulateur);
-        double TotalPayeParCommande(int idCommande);
-        double TotalValideParCommande(int idCommande);
+        decimal TotalPayeParCommande(int idCommande);
+        decimal TotalValideParCommande(int idCommande);
         string GenererNumeroRecu();
     }
 }
